Rec Facial y BD/requirements.txt
face_recognition
opencv-python==4.6.0.66
numpy
albumentations
supabase
matplotlib
python-dotenv
promisio
